+++ R/get_grids.R
@@ -74,16 +74,7 @@
 #' head(fit)
 #' }
 get_grids <- function(trimtonOut, Grid) {
-  len1 <- dim(trimtonOut)[2]
-  childli <- childnew <- rectidli <- vector("list", len1)
-  for (i in 1:len1) {
-    childli[[i]] <- trimtonOut[, i]
-  }
-  for (u in 1:len1) {
-    rectidli[[u]] <- which(childli[[u]] == 1, arr.ind = TRUE)
-  }
-  for (z in 1:len1) {
-    childnew[[z]] <- Grid[rectidli[[z]], ]
-  }
-  return(childnew)
+  lapply(seq_len(ncol(trimtonOut)), function(i) {
+    Grid[trimtonOut[, i] == 1, , drop = FALSE]
+  })
 }